src/: Support IPv6 addresses in `sortlist'.
This involves a fair amount of messing about.  There's a new file
`addrfam.c' containing a number of address-family-specific facts and
utilities, packaged together in a handy structure.  The `sortlist'
representation now attaches one of these to each entry, and now keeps
a union of possible address types.

Lookup of addresses in the sortlist is also modified, to check the
address family before going any further.

Signed-off-by: Mark Wooding <mdw@distorted.org.uk>