Change how addr queries track which subqueries to make.
Now we have a vector of query types we might want to make, and a pair of
bitmaps tracking whether (a) we want to issue that kind of query,
and (b) whether we've had an answer to it yet.  This replaces
the (sometimes tricky) messing about with raw rrtype vectors.

As a bonus, these bitmaps are now set in the context structure, where
they can be pushed through adns__internal_submit and into qs_addr, so we
no longer need adns__qf_nosend.  This is, accordingly, killed.