src/types.c: Use the new `checklabel' machinery for PTR queries.
Previously we'd check the query name and parse it into an address while
parsing the answer to the initial query, which means that you only get
an error about the name being malformed after waiting for the answer.
Now we have a better placed hook for dealing with this, so we should use
it.