Bodge adns_r_addr query-id selection and fix regression tests to match.
A future change is going to change how query-ids are allocated for
adns_r_addr queries.  So here we apply a simple bodge which will produce
the same allocation pattern, but without the complexity of the rest of
the change, and fix the test suite to match.

I used the script regress/hack-query-ids to help me modify the
case-*.sys files, though this was primarily a rather tedious manual
process.

Signed-off-by: Mark Wooding <mdw@distorted.org.uk>