src/types.c: Better record parsing for addr queries.
The parser now needs to know which kind of record it's meant to be
dealing with, so that it can convert it into the right kind of socket
address.  As a result, pap_findaddrs needs to pass in the right record
type when parsing addresses from the additional section, so we separate
out the main work into a separate function.

Signed-off-by: Mark Wooding <mdw@distorted.org.uk>