regress: Introduce hnonfuzz.c and Ttestinputfd
We are going to want to build a fuzzer target binary which will be
very like a playback binary, but provides a bit more wrapping.

For now we introduce hnonfuzz.o and link it into all the
currently-generated executables.

To make it not be an empty translation unit, combine this with:

Break out Ttestinputfd from hplayback.c.m4.

Neither of these changes produce any functional change.  (Ttestinputfd
ends up linked into *_record as well as *_playback but this is of no
consequence.)

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>