timeout robustness: Track start time and duration
This allows us to detect if the clock rewinds, and restart our
timeouts from the new time.  Otherwise we might try to wait a very
long time.

The result is that clock instability may now produce spurious failures
of some queries, but it should no longer cause queries to be
indefinitely delayed.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>